requirements.txt
praw~=7.7
beautifulsoup4~=4.12
requests~=2.28
Pillow~=9.5
PyYAML~=6.0
sqlalchemy~=1.4
voluptuous~=0.13